services/llm-worker/requirements.txt
# LLM
litellm==1.30.0

# NATS
nats-py==2.7.0

# Data handling
pydantic==2.6.0
pydantic-settings==2.2.0

# Observability
prometheus-client==0.20.0
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-exporter-otlp==1.23.0
structlog==24.1.0

# Utilities
python-dotenv==1.0.1
tenacity==8.2.3
tiktoken==0.6.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-asyncio==0.23.0
